change type of length parameter to int.

+++ cmd/storegen/wrapper/template.go
@@ -402,8 +402,8 @@ func (s* Storage_{{.Name}}) Length() (int) {
 	return int(rv.Big().Int64())
 }
 
-func (s* Storage_{{.Name}}) Resize(length uint64) {	
-	s.db.SetState(s.addr, common.BigToHash(s.slot), common.BigToHash(big.NewInt(0).SetUint64(length)))
+func (s* Storage_{{.Name}}) Resize(length int) {	
+	s.db.SetState(s.addr, common.BigToHash(s.slot), common.BigToHash(big.NewInt(0).SetUint64(uint64(length))))
 
 	slice := make([]*{{$elem.Type}}, length, length+50)
 	copy(slice, *s.obj)
@@ -414,7 +414,7 @@ func (s* Storage_{{.Name}}) Get(index int) ( *Storage_{{$elem.Type}} ) {
 	// Value: {{ printf "%#v" $elem }}
 	length := s.Length()
 	if length < index {
-		s.Resize(uint64(index+1))
+		s.Resize(index+1)
 	}
 
 	hash := crypto.Keccak256Hash(common.BigToHash(s.slot).Bytes())
